add refresh function to library
